refactor(gh): extract gh/comments submodule
Second of the 5-PR series splitting src/gh/mod.rs into per-resource
submodules (issue #301). Moves the gist-comment fetch/parse/pagination
cluster into gh/comments.rs: COMMENTS_PAGE_SIZE, last_page,
parse_link_rel (plus its private http_headers_section/http_body_section
helpers), comments_total_from_probe, gist_comments_probe_plan/
gist_comments_page_plan, fetch_gist_comments_probe(_with)/
fetch_gist_comments_page(_with), parse_gist_comments_json, and
parse_gist_comment_counts, along with their unit tests.

parse_gh_gists and GhCommentUser stay in gh/mod.rs as shared kernel
(also used by the not-yet-extracted gists/forks/revisions clusters);
comments.rs reaches them via `use super::{parse_gh_gists, GhCommentUser};`,
matching gh/stars.rs's existing `use super::X` convention for kernel
access. gh/mod.rs re-exports comments' public items by explicit name,
so none of the existing crate::gh::X call sites need to change.

No signature or behavior change.

Part of #301
